--- control-with-dropdown/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "control-with-url",
-  "version": "1.0.0",
-  "description": "",
-  "main": "app.js",
-  "scripts": {
-    "start": "node app.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "axios": "^0.19.0",
-    "express": "^4.17.1"
-  }
-}
