--- control-with-dropdown/public/index.html
@@ -1,34 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <head>
-      <title>IFrame examples</title>
-      <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-      <link rel="stylesheet" type="text/css" href="style.css">
-    </head>
-    <body>
-      <div class="container center">
-        <div class="card center mt-5" >
-          <div class="card-header h5">Page Filter</div>
-          <div class="card-img-top">
-            <iframe id="pivot" src="http://localhost:9095/pivot/d/druid_wikipedia"></iframe>
-          </div>
-          <div class ="card-body m1">
-            <div class="dropdown">
-              <button class="btn btn-primary dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown">
-               Dimension
-              </button>
-              <div class="dropdown-menu">
-                <a class="dropdown-item" id="Page" onclick="setDimension(id)">Page</a>
-                <a class="dropdown-item" id="User" onclick="setDimension(id)">User</a>
-                <a class="dropdown-item" id="Comment" onclick="setDimension(id)">Comment</a>
-              </div>
-          </div>
-         </div>
-        <script src="logic.js"></script>
-        <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
-        <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script>
-        </div>
-       </div>
-    </body>
-</html>
