--- control-with-dropdown/public/style.css
@@ -1,8 +1,0 @@
-* {
-  font-family: Arial;
-}
-
-iframe {
-  width: 100%;
-  height: 500px;
-}
